refactor(sandbox): split credential_firewall's test module into a sibling file
thermo-nuclear-code-quality-review ruling: a 1259-line file (437-line
implementation + ~820-line inline test module) is a real maintainability
smell under the repo's file-size threshold (.claude/rules/architecture.md).
Move `mod tests` to sandbox_process/credential_firewall/tests.rs — no
behavior change, no new abstraction, same 20 tests still pass. The
implementation file itself is now 489 lines.
